+++ src/main/java/ak2/services/ak2/common/view/Ak2RedirectView.java
@@ -1,0 +1,11 @@
+package ak2.services.ak2.common.view;
+
+import org.springframework.web.servlet.view.RedirectView;
+
+public class Ak2RedirectView extends RedirectView {
+
+  public Ak2RedirectView(String path) {
+    super("http://ak2.prod.jp.local" + path);
+  }
+
+}
