--- src/main/java/ak2/services/ak2/web/dto/SessionForm.java
@@ -1,14 +1,0 @@
-package ak2.services.ak2.web.dto;
-
-import lombok.Data;
-
-import java.io.Serial;
-import java.io.Serializable;
-
-@Data
-public class SessionForm implements Serializable {
-    @Serial
-    private static final long serialVersionUID = 1L;
-
-    public String sessionValue = "No Data was found";
-}
