--- src/main/java/ak2/services/ak2/web/enums/MessageEnum.java
@@ -1,5 +1,0 @@
-package ak2.services.ak2.web.enums;
-
-public enum MessageEnum {
-  
-}
