--- src/main/java/ak2/services/ak2/web/enums/SessionEnum.java
@@ -1,15 +1,0 @@
-package ak2.services.ak2.web.enums;
-
-public enum SessionEnum {
-    SESSION_ENUM("session_key");
-
-    private final String key;
-
-    SessionEnum(String key) {
-        this.key = key;
-    }
-
-    public String getKey() {
-        return key;
-    }
-}
