+++ src/main/java/ak2/services/ak2/web/session/dto/SessionForm.java
@@ -1,0 +1,14 @@
+package ak2.services.ak2.web.session.dto;
+
+import lombok.Data;
+
+import java.io.Serial;
+import java.io.Serializable;
+
+@Data
+public class SessionForm implements Serializable {
+  @Serial
+  private static final long serialVersionUID = 1L;
+
+  public String sessionValue = "No Data was found";
+}
