+++ src/main/java/ak2/services/ak2/web/session/enums/MessageEnum.java
@@ -1,0 +1,5 @@
+package ak2.services.ak2.web.session.enums;
+
+public enum MessageEnum {
+
+}
